requirements.txt
click==8.1.7
Flask==2.0.1
flask-cors==3.0.10
gunicorn==20.0.4
itsdangerous==2.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dotenv==0.19.0
recipe-scrapers==15.0.0
ingredient-parser-nlp==1.0.0
Werkzeug==2.0.1
